HBX-1599: Extract interface 'org.hibernate.tool.api.export.ArtifactCollector' from 'org.hibernate.tool.internal.export.DefaultArtifactCollector'
