+++ orm/src/main/java/org/hibernate/tool/hbm2x/Exporter.java
@@ -7,8 +7,8 @@ import java.io.File;
 import java.util.Properties;
 
 import org.hibernate.boot.Metadata;
+import org.hibernate.tool.api.export.ArtifactCollector;
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
-import org.hibernate.tool.internal.exporter.DefaultArtifactCollector;
 
 /**
  * @author max and david
@@ -44,13 +44,13 @@ public interface Exporter {
 	 * 
 	 * @param collector Instance to be consulted when adding a new file.
 	 */
-	public void setArtifactCollector(DefaultArtifactCollector collector);
+	public void setArtifactCollector(org.hibernate.tool.api.export.ArtifactCollector collector);
 	
 	/**
 	 * 
 	 * @return artifact collector
 	 */
-	public DefaultArtifactCollector getArtifactCollector();
+	public ArtifactCollector getArtifactCollector();
 	
 	/**
 	 * Called when exporter should start generating its output
